fixed aceLow glitch
fixed a glitch where setting an ace to be 1 instead of 11 would permanently change it, rather than just for that blackjack game. if the card came up again in a future game, it would still have value 1. this glitch has been fixed by resetting all the cards in Deck.shuffle()